// frontend/src/api/generated/model/countUnit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type CountUnit = typeof CountUnit[keyof typeof CountUnit];


export const CountUnit = {
  Undefined: 'Undefined',
  AttackedPlantPercentage: 'AttackedPlantPercentage',
  AttackedPlantsPer10MetreRow: 'AttackedPlantsPer10MetreRow',
  InsectsPerPlant: 'InsectsPerPlant',
  InsectsPerTenPlants: 'InsectsPerTenPlants',
  InsectsPerTrap: 'InsectsPerTrap',
  LesionedLeafAreaPercentage: 'LesionedLeafAreaPercentage',
  SeverityScore1To9: 'SeverityScore1To9',
  Presence: 'Presence',
} as const;