// frontend/src/api/generated/model/crop.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type Crop = typeof Crop[keyof typeof Crop];


export const Crop = {
  Undefined: 'Undefined',
  Soybean: 'Soybean',
  Corn: 'Corn',
  Cotton: 'Cotton',
  Coffee: 'Coffee',
  Sugarcane: 'Sugarcane',
  Wheat: 'Wheat',
  Beans: 'Beans',
  Rice: 'Rice',
  Sorghum: 'Sorghum',
  Other: 'Other',
} as const;